requirements.txt
torch
numpy
Pillow
moviepy
